voice_memo_assistant/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bot"]

[project]
name = "voice_memo_assistant"
version = "0.1.0"
description = "Telegram bot that transcribes voice messages, summarizes them with LLM, and stores in a database"
requires-python = ">=3.10"
authors = [
    {name = "VoiceMemo Assistant Team"}
]
dependencies = [
    "python-telegram-bot>=20.8",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "vosk>=0.3.45",
    "soundfile>=0.12.1",
    "resampy>=0.4.2",
    "openai>=1.3.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0",
    "pytest>=7.4.2"
]

[project.scripts]
voice-memo-bot = "bot.main:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88 